moving helper functions
